requirements.txt
PySimpleGuiQT~=0.35.0
PySimpleGui
Image
Pillow
opencv-python
numpy
matplotlib
numpy